requirements.txt
streamlit==1.28.1
groq==0.11.1
openai==1.12.0
google-api-python-client==2.108.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
pandas==2.1.4
python-dotenv==1.0.0
pydantic>=2.0.0,<3.0.0
